refactor: move async cache logic from FeatureRepository to EnhancedFeatureRepository
